add comments and flowchart

+++ SquareRootChecker.java
@@ -1,3 +1,11 @@
+/**
+*
+* Reads in scores until a negative value is encountered, at which point it determines and prints the average
+*
+* @author Jack Navin-Weinstein
+* @version Thursday,16 October
+*
+*/
 import java.util.Scanner;
 
 public class SquareRootChecker{
@@ -6,7 +14,7 @@ public class SquareRootChecker{
    Scanner scnr = new Scanner(System.in);
    
    System.out.println("Enter positve interger: ");
-   int number = scnr.nextInt();
+   double number = scnr.nextDouble();
    
    if (number<0){
    System.out.println("Enter zero or quantity greater than 0: ");
